MTR-44 Player Minimap on lookdown
Changed the Look Limit on the X axis to encompass looking to the top of the player cube where the minimap is. Currently is a live raw image casted from an overview camera, so lacks the ability to remove fog of war upon exploration. Hoping to be able to set waypoint markers using the minimap